updates to the UML model 2024-09-10
hello @kwkwan i have updated the content of the model for both building and landuse. if you could update that for the branch that would be very helpful.

i have added in an actual image now to see if that content can be placed with in a cell.

thank you