updates to UML model export
hello @kwkwan i have made updates in the uml model, please use this .xmi file. additionally i have changed the file name of the .xmi and updated that in the document, so we dont have to change it all the time. i can manage the uploaded file name locally.  i will upload the commit but stop the job. please finalize the updates for me. thank you.