change of edition to v5 for both source documents
work on v4-1 was completed. we wanted to keep those folders in sources but we are now working on the v5 folders for documents 001-v5 and 002-v5;

collection.yml and metanorma.yml in the project root were also edited

document.adoc (in both 001-v5 and 002-v5) were also revised;

revdate, published-date and copyright-year will need to be checked;

i will check the result in the morning.