update document.adoc to remove novalid
doing testing with Nick at the end of Nov 2025 to reinstate the output of the document.err.html files the fix involved removing the :novalid: statement in the document.adoc file. this is now removed in both documents sources/001-v5, sources/002-v5;

middle-dots in sources/002-v5 were also removed on two lines to unify with document 001-v5.